refactor(snapshots): move to separate module
