+++ README.md
@@ -74,7 +74,7 @@ pip install "modflow-devtools[test]"
 
 To install from source and set up a development environment please see the [developer documentation](DEVELOPER.md).
 
-To import `pytest` fixtures in a project consuming `modflow-devtools`, add the following to a `conftest.py` file:
+To import `pytest` fixtures in a project consuming `modflow-devtools`, add the following to a test file or `conftest.py` file:
 
 ```python
 pytest_plugins = [ "modflow_devtools.fixtures" ]
