+++ modflow_devtools/fixtures.py
@@ -1,5 +1,4 @@
 from collections import OrderedDict
-from io import BytesIO, StringIO
 from itertools import groupby
 from os import PathLike, environ
 from pathlib import Path
@@ -9,91 +8,7 @@ from typing import Dict, Generator, List, Optional
 from modflow_devtools.imports import import_optional_dependency
 from modflow_devtools.misc import get_namefile_paths, get_packages
 
-np = import_optional_dependency("numpy")
 pytest = import_optional_dependency("pytest")
-syrupy = import_optional_dependency("syrupy")
-
-# ruff: noqa: E402
-from syrupy.extensions.single_file import (
-    SingleFileSnapshotExtension,
-    WriteMode,
-)
-from syrupy.types import (
-    PropertyFilter,
-    PropertyMatcher,
-    SerializableData,
-    SerializedData,
-)
-
-# snapshot extensions
-
-
-class BinaryArrayExtension(SingleFileSnapshotExtension):
-    """
-    Binary snapshot of a NumPy array. Can be read back into NumPy with
-    .load(), preserving dtype and shape. This is the recommended array
-    snapshot approach if human-readability is not a necessity, as disk
-    space is minimized.
-    """
-
-    _write_mode = WriteMode.BINARY
-    _file_extension = "npy"
-
-    def serialize(
-        self,
-        data,
-        *,
-        exclude=None,
-        include=None,
-        matcher=None,
-    ):
-        buffer = BytesIO()
-        np.save(buffer, data)
-        return buffer.getvalue()
-
-
-class TextArrayExtension(SingleFileSnapshotExtension):
-    """
-    Text snapshot of a NumPy array. Flattens the array before writing.
-    Can be read back into NumPy with .loadtxt() assuming you know the
-    shape of the expected data and subsequently reshape it if needed.
-    """
-
-    _write_mode = WriteMode.TEXT
-    _file_extension = "txt"
-
-    def serialize(
-        self,
-        data: "SerializableData",
-        *,
-        exclude: Optional["PropertyFilter"] = None,
-        include: Optional["PropertyFilter"] = None,
-        matcher: Optional["PropertyMatcher"] = None,
-    ) -> "SerializedData":
-        buffer = StringIO()
-        np.savetxt(buffer, data.ravel())
-        return buffer.getvalue()
-
-
-class ReadableArrayExtension(SingleFileSnapshotExtension):
-    """
-    Human-readable snapshot of a NumPy array. Preserves array shape
-    at the expense of possible loss of precision (default 8 places)
-    and more difficulty loading into NumPy than TextArrayExtension.
-    """
-
-    _write_mode = WriteMode.TEXT
-    _file_extension = "txt"
-
-    def serialize(
-        self,
-        data: "SerializableData",
-        *,
-        exclude: Optional["PropertyFilter"] = None,
-        include: Optional["PropertyFilter"] = None,
-        matcher: Optional["PropertyMatcher"] = None,
-    ) -> "SerializedData":
-        return np.array2string(data, threshold=np.inf)
 
 
 # fixtures
@@ -174,21 +89,6 @@ def tabular(request) -> str:
     if tab not in ["raw", "recarray", "dataframe"]:
         raise ValueError(f"Unsupported value for --tabular: {tab}")
     return tab
-
-
-@pytest.fixture
-def array_snapshot(snapshot):
-    return snapshot.use_extension(BinaryArrayExtension)
-
-
-@pytest.fixture
-def text_array_snapshot(snapshot):
-    return snapshot.use_extension(TextArrayExtension)
-
-
-@pytest.fixture
-def readable_array_snapshot(snapshot):
-    return snapshot.use_extension(ReadableArrayExtension)
 
 
 # configuration hooks
